pyproject.toml
[tool.poetry]
name = "sanic_test"
version = "0.1.0"
description = ""
authors = ["Satdinov <satdinov.rsh@yandex.ru>"]

[tool.poetry.dependencies]
python = ">=3.7.2,<3.11"
toml = "^0.10"
sentry-sdk = "^1.5"
sanic = {version = "^21.12", extras = ["ext"]}
sanic-jwt = "^1.7"
pydantic = {version = "^1.8", extras = ["email"]}
argon2-cffi = "^21.1"
aioredis = "^2.0"
gino = "^1.0"
gino-sanic = {git = "https://github.com/sandlabs/gino-sanic.git", branch = "master"}
alembic = "^1.6"
psycopg2-binary = "^2.8"
aio-pika = "^6.8"
pyotp = "^2.6.0"
hexbytes = "^0.2.2"
eth-account = "^0.6.1"
Jinja2 = "^3.1.1"
pytest = "^6.2"
pytest-asyncio = "^0.16"
sanic-testing = "^22.3"
pytest-postgresql = "^4.0"
sanic-ext = "^22.3.1"

[tool.isort]
line_length = 120
multi_line_output = 5
lines_after_imports = 2
atomic = true
combine_star = true
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
skip = [".git", ".env", ".venv", "env", "venv", "node_modules", "alembic"]

[tool.poetry.dev-dependencies]
prospector = "^1.7.7"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
